+++ HISTORY.md
@@ -1,3 +1,16 @@
+### v0.0.5
+* `Text` element now supported setting `alignment` props (left|center|right) to set text alignment.
+
+* `transform` and `projection` prop on `<Spider />` component will effect whole graph but node only.
+
+* to set `transform` or `projection` prop of node, set `transform` or `projection` prop of `<Node />` component directly, 
+
+  or set `NodeTransform` or `NodeProjection` attribute of `<Spider />` 
+
+
+### v0.0.4
+using tidier drawing algorithm to draw tree
+
 ### v0.0.3
 
 tree-layout supported.